add env to gitignore and decimals to emp salary

+++ schema.sql
@@ -16,7 +16,7 @@ CREATE TABLE roles (
     job_title VARCHAR(30) NOT NULL,
     id INT NOT NULL AUTO_INCREMENT,
     dept_id INT NOT NULL,
-    salary INT NOT NULL,
+    salary DECIMAL(7,2) NOT NULL,
     PRIMARY KEY(id) 
 );
 
@@ -27,7 +27,7 @@ CREATE TABLE employees (
     last_name VARCHAR(30) NOT NULL,
     job_title VARCHAR(30) NOT NULL,
     dept_id INT NOT NULL,
-    salary INT NOT NULL,
+    salary DECIMAL(7,2) NOT NULL,
     managers_id INT NOT NULL,
     PRIMARY KEY(id)
 );
